Revert "server.c: Fix warning the use of `mktemp' is dangerous"
This is obviously breaking the unix domain socket code here. We
only want a filename and we try to bind on this. So like the comment
in the patch says. If someone is guessing our random name and is
creating a socket our bind is going to fail and we will retry.

This reverts commit eeb603239c1ad561b95e26481283a837790d9710.