main.c: Fix a crash when clearing a profile
In the case the user has generated a profile and then
is pressing the swipe button we will crash. This is because
the selection_changed callback will be called and we try
to access the profile and then will crash.

Change the code to check for n_bytes only after we have
checked if there is a selection. This is fixing the crash.